[MouseWithoutBorders] - moving Common.Launch.cs -> Core\Launch.cs - #35155
